agent/pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.2",
    "openai>=2.14.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]
